spec: move cgo stuff to their own file
so it can build without cgo since seccomp requires it.
